+++ src/Request/Highlight.php
@@ -37,6 +37,7 @@ class Highlight extends RequestCollection
             ->addParam('phone_id', $this->ig->phone_id)
             ->addParam('battery_level', $this->ig->getBatteryLevel())
             ->addParam('is_charging', $this->ig->getIsDeviceCharging())
+            ->addParam('is_dark_mode', (int) $this->ig->getIsDarkModeEnabled())
             ->addParam('will_sound_on', (int) $this->ig->getSoundEnabled())
             ->getResponse(new Response\HighlightFeedResponse());
     }